Move UmbralParameters class to params.py
